Enhanced UX and information access, Power class added
*Card:
- Added enum class CID to list all cards in the game.
- Added member CID id to Card class, along with getter method.
- Refactored display method to now display energy using Unicode, to reduce risk of confusion with position using to_energy() helper.

*Pile:
- Added IndexedCard struct
- Added sortBy method to allow displaying copies of piles sorted by different categories
- added indexed method to obfuscate order of pile if needed.

*Character:
- Changed mutation methods to now return signed integers. and return real mutation values instead of expected values.

*Player:
- Changed transferCards methods to return unsigned integers to indicate how many cards were really transferred instead of the expected value.
- displayPlayerPile method now uses indexed to obfuscate pile order if needed.

*Effect:
- Changed EffectReport to signed integers
- log method is now more accurate when it comes to reporting actual mutation values by using EffectReport.

***Power:
- Added new power class to represent buff/debuffs. Added rudimentary stacking logic. Pending linking Power class to rest of codebase to fully integrate buffs and debuffs.

*RNG:
- nextInt now returns uint64_t instead of int, which was causing range issues.

*UI:
- changed Color::common to improve readability.
- Added to_energy method to use a look up table to render a card's energy cost

+++ include/game_logic/rng.h
@@ -13,7 +13,7 @@ class RNG{
     RNG(uint64_t s);
 
     //Returns a random integer between min and max, inclusive.
-    int nextInt(int min, int max);
+    uint64_t nextInt(int min, int max);
     //Returns a random float between min and max.
     float nextFloat(float min, float max);
